fix(ppu): fix ClearBackdrop buffer overrun; widen widescreen to uint16 (32:9)
The widescreen cap was stuck at "152" because ClearBackdrop looped
`for (i=0; i != countof; i += 4)` (4 entries/step), which only terminates when
kPpuXPixels = 256+2*extra is a multiple of 4 — i.e. an even cap. Odd caps ran the
loop past bgBuffers/objBuffer into the audio/DSP heap, surfacing as a crash in
dsp_saveload on save-state load. Fixed to handle any buffer size.

With that gone, widen for true ultrawide:
- extended_aspect_ratio + Ppu.extraLeftRight/LeftCur/RightCur: uint8 -> uint16
- kPpuExtraLeftRight 152 -> 384 (1024px = full big-area width, ~4.3:1, covers 32:9)
- WasmGetViewportInfo / render.ts: pack those fields as uint16
- aspect-ratio.ts MAX_EXTRA 152 -> 384

Verified true 32:9 renders clean in Kakariko; no regression at normal ratios.

+++ core/zelda3/src/config.h
@@ -59,7 +59,7 @@ typedef struct Config {
   uint8 audio_channels;
   uint16 audio_samples;
   bool autosave;
-  uint8 extended_aspect_ratio;
+  uint16 extended_aspect_ratio;  // extra columns per side; uint16 so it can exceed 255 (>3.19:1)
   bool extend_y;
   bool no_sprite_limits;
   bool display_perf_title;
